+++ crypto-ffi/src/generic/e2ei_context.rs
@@ -10,7 +10,7 @@ use super::{
 
 #[uniffi::export]
 impl CoreCryptoContext {
-    /// See [core_crypto::mls::context::CentralContext::e2ei_new_enrollment]
+    /// See [core_crypto::context::CentralContext::e2ei_new_enrollment]
     pub async fn e2ei_new_enrollment(
         &self,
         client_id: String,
@@ -36,7 +36,7 @@ impl CoreCryptoContext {
             .map(E2eiEnrollment)?)
     }
 
-    /// See [core_crypto::mls::context::CentralContext::e2ei_new_activation_enrollment]
+    /// See [core_crypto::context::CentralContext::e2ei_new_activation_enrollment]
     pub async fn e2ei_new_activation_enrollment(
         &self,
         display_name: String,
@@ -54,7 +54,7 @@ impl CoreCryptoContext {
             .map(E2eiEnrollment)?)
     }
 
-    /// See [core_crypto::mls::context::CentralContext::e2ei_new_rotate_enrollment]
+    /// See [core_crypto::context::CentralContext::e2ei_new_rotate_enrollment]
     pub async fn e2ei_new_rotate_enrollment(
         &self,
         display_name: Option<String>,
@@ -72,13 +72,13 @@ impl CoreCryptoContext {
             .map(E2eiEnrollment)?)
     }
 
-    /// See [core_crypto::mls::context::CentralContext::e2ei_register_acme_ca]
+    /// See [core_crypto::context::CentralContext::e2ei_register_acme_ca]
     pub async fn e2ei_register_acme_ca(&self, trust_anchor_pem: String) -> CoreCryptoResult<()> {
         self.context.e2ei_register_acme_ca(trust_anchor_pem).await?;
         Ok(())
     }
 
-    /// See [core_crypto::mls::context::CentralContext::e2ei_register_intermediate_ca_pem]
+    /// See [core_crypto::context::CentralContext::e2ei_register_intermediate_ca_pem]
     pub async fn e2ei_register_intermediate_ca(&self, cert_pem: String) -> CoreCryptoResult<Option<Vec<String>>> {
         Ok(self
             .context
@@ -87,12 +87,12 @@ impl CoreCryptoContext {
             .map(Into::into)?)
     }
 
-    /// See [core_crypto::mls::context::CentralContext::e2ei_register_crl]
+    /// See [core_crypto::context::CentralContext::e2ei_register_crl]
     pub async fn e2ei_register_crl(&self, crl_dp: String, crl_der: Vec<u8>) -> CoreCryptoResult<CrlRegistration> {
         Ok(self.context.e2ei_register_crl(crl_dp, crl_der).await.map(Into::into)?)
     }
 
-    /// See [core_crypto::mls::context::CentralContext::e2ei_mls_init_only]
+    /// See [core_crypto::context::CentralContext::e2ei_mls_init_only]
     pub async fn e2ei_mls_init_only(
         &self,
         enrollment: std::sync::Arc<E2eiEnrollment>,
@@ -115,7 +115,7 @@ impl CoreCryptoContext {
             .map(Into::into)?)
     }
 
-    /// See [core_crypto::mls::context::CentralContext::e2ei_rotate_all]
+    /// See [core_crypto::context::CentralContext::e2ei_rotate_all]
     pub async fn e2ei_rotate_all(
         &self,
         enrollment: std::sync::Arc<E2eiEnrollment>,
@@ -132,7 +132,7 @@ impl CoreCryptoContext {
             .try_into()
     }
 
-    /// See [core_crypto::mls::context::CentralContext::e2ei_enrollment_stash]
+    /// See [core_crypto::context::CentralContext::e2ei_enrollment_stash]
     pub async fn e2ei_enrollment_stash(&self, enrollment: std::sync::Arc<E2eiEnrollment>) -> CoreCryptoResult<Vec<u8>> {
         let enrollment = std::sync::Arc::into_inner(enrollment).ok_or_else(|| CryptoError::LockPoisonError)?;
         let enrollment = std::sync::Arc::into_inner(enrollment.0)
@@ -142,7 +142,7 @@ impl CoreCryptoContext {
         Ok(self.context.e2ei_enrollment_stash(enrollment).await?)
     }
 
-    /// See [core_crypto::mls::context::CentralContext::e2ei_enrollment_stash_pop]
+    /// See [core_crypto::context::CentralContext::e2ei_enrollment_stash_pop]
     pub async fn e2ei_enrollment_stash_pop(&self, handle: Vec<u8>) -> CoreCryptoResult<E2eiEnrollment> {
         Ok(self
             .context
@@ -153,7 +153,7 @@ impl CoreCryptoContext {
             .map(E2eiEnrollment)?)
     }
 
-    /// See [core_crypto::mls::context::CentralContext::e2ei_conversation_state]
+    /// See [core_crypto::context::CentralContext::e2ei_conversation_state]
     pub async fn e2ei_conversation_state(&self, conversation_id: Vec<u8>) -> CoreCryptoResult<E2eiConversationState> {
         Ok(self
             .context
