+++ crypto/src/e2e_identity/rotate.rs
@@ -10,7 +10,7 @@ use mls_crypto_provider::TransactionalCryptoProvider;
 use crate::e2e_identity::init_certificates::NewCrlDistributionPoint;
 #[cfg(not(target_family = "wasm"))]
 use crate::e2e_identity::refresh_token::RefreshToken;
-use crate::mls::context::CentralContext;
+use crate::context::CentralContext;
 use crate::{
     mls::credential::{ext::CredentialExt, x509::CertificatePrivateKey, CredentialBundle},
     prelude::{
