+++ crypto/src/mls/conversation/db_count.rs
@@ -1,4 +1,4 @@
-use crate::{mls::context::CentralContext, prelude::MlsCentral};
+use crate::prelude::MlsCentral;
 use core_crypto_keystore::{
     connection::FetchFromDatabase,
     entities::{
@@ -7,6 +7,7 @@ use core_crypto_keystore::{
         PersistedMlsPendingGroup,
     },
 };
+use crate::context::CentralContext;
 
 #[derive(Debug, Clone, Eq, PartialEq)]
 pub struct EntitiesCount {
