+++ crypto/src/mls/credential/crl.rs
@@ -1,4 +1,4 @@
-use crate::{e2e_identity::init_certificates::NewCrlDistributionPoint, mls::context::CentralContext};
+use crate::e2e_identity::init_certificates::NewCrlDistributionPoint;
 use crate::{CryptoError, CryptoResult};
 use core_crypto_keystore::{connection::FetchFromDatabase, entities::E2eiCrl};
 use mls_crypto_provider::TransactionalCryptoProvider;
@@ -9,6 +9,7 @@ use openmls::{
 use openmls_traits::OpenMlsCryptoProvider;
 use std::collections::HashSet;
 use wire_e2e_identity::prelude::x509::extract_crl_uris;
+use crate::context::CentralContext;
 
 #[cfg_attr(not(test), tracing::instrument(err, skip_all))]
 pub(crate) fn extract_crl_uris_from_credentials<'a>(
